# Rework read_bytes to accept mutable buffer