By moving `read_bytes` to accept a mutable buffer:

- Remove all allocations (hidden or otherwise) from the API
- Increase performance (aligned benchmark improved from 12.6us to 4.3us
and unaligned improved from 29.8us to 17.8us). I was kinda shocked, I
would have thought that creating a borrowed cow of slice data would be
cheaper than copying, but it looks like that is not the case.

This is a breaking change, but hopefully it gives users a chance to
reuse allocations for performance and space improvements.

Instead of

```rust
let mut bitter = BitGet::new(&[0b1010_1010, 0b0101_0101]);
assert_eq!(
    bitter.read_bytes(2).map(|x| x.into_owned()),
    Some(vec![0b1010_1010, 0b0101_0101])
)
```

write:

```rust
let mut buf = [0u8; 2];
let mut bitter = BitGet::new(&[0b1010_1010, 0b0101_0101]);
assert!(bitter.read_bytes(&mut buf));
assert_eq!(&buf, &[0b1010_1010, 0b0101_0101]);
```

The new API should be more reminiscent of the Read trait's read method,
which also takes a mutable buffer, but instead of returning a result of
bytes read we know that either we'll fill the entire buffer or there
isn't enough data.